mangacan update collector

+++ model/episode.go
@@ -4,7 +4,7 @@ import "time"
 
 type Episode struct {
 	ID        int
-	No        float32
+	No        float64
 	Name      string
 	CreatedAt time.Time
 	UpdatedAt time.Time
